# HardwareSerial: don't call STM32 HAL for every bytes to transfer