--- a/cores/arduino/stm32/uart.h
+++ b/cores/arduino/stm32/uart.h
@@ -80,6 +80,7 @@ struct serial_s {
   uint16_t tx_head;
   volatile uint16_t rx_head;
   volatile uint16_t tx_tail;
+  size_t tx_size;
 };
 
 /* Exported constants --------------------------------------------------------*/
@@ -186,7 +187,7 @@ void uart_config_lowpower(serial_t *obj);
 size_t uart_write(serial_t *obj, uint8_t data, uint16_t size);
 int uart_getc(serial_t *obj, unsigned char *c);
 void uart_attach_rx_callback(serial_t *obj, void (*callback)(serial_t *));
-void uart_attach_tx_callback(serial_t *obj, int (*callback)(serial_t *));
+void uart_attach_tx_callback(serial_t *obj, int (*callback)(serial_t *), size_t size);
 
 uint8_t serial_tx_active(serial_t *obj);
 uint8_t serial_rx_active(serial_t *obj);