**Summary**

HardwareSerial: don't call STM32 HAL for every bytes to transfer

Instead of calling HAL for each and every bytes,
try to call it only once for a string.
Several call may be needed to manage internal circular buffer (tx_buff)

Fixes #1309