@ABOSTM 2021-03-11
**Performances measurements:**
Note: This PR subserve, by implementation, sending multi char string instead of single char.
This is the major usage.

Tests performed with UART @115200

| Test | Before PR | After PR| Performance|
| --- | --- | --- | --- |
| Nucleo_l476RG   <br> sending string of 20 char | 1,80 ms  | 1,72 ms  | gain: 4,4% |
| Nucleo_l476RG  <br>  sending 20 char one by one | 1,80 ms  | 1,88 ms  | loss: 4,4% |
| BluePill       <br>          sending string of 20 char | 1,90 ms  | 1,73 ms  | gain: 8,9% |
| BluePill       <br>          sending 20 char one by one | 1,90 ms  | 1,90 ms  | stable |